pyproject.toml
[tool.poetry]
name = "dsp"
version = "0.1.0"
description = "Dyson Sphere Program Calculations and Visualizations"
authors = ["Will Barden <wfbarden@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyvis = "0.3.1"
networkx = "^3.1"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
